Make pm and post_mortem easier to use by using sys.exec_info() if sys.last_traceback hasn't been set.
fns.py: get_last_tb_or_frame_tb(frameno=1):

Intended to be used going into post mortem routines.  If
sys.last_traceback is set, we will return that and assume that this is
what post-mortem will want. If sys.last_traceback has not been set,
then perhaps we *about* to raise an error and are fielding an
exception. So assume that sys.exec_info()[frameno] is where we want to
look.