Patch from Matt Fleming:
Had checked to see whether we were set to 'stop' on a signal inside
trace_dispatch and only then, if the signal handler was not our custom
one, reset the signal handler to our custom handler and saved the old
handler (which could be a handler from a user's program) in
old_handler.

Now, we check to see if the signal handler has been replaced by
something that is not:

 signal.SIG_DFL - the default handler in Python
 signal.SIG_IGN - sets python to ignore the signal
 sighandler.handle - pydb's signal handler

If this is so, then it must have been set to a signal handler from a
user's file. In which case we store it, and use it later if the
programmer controlling pydb decides they want to pass the signal to
their program.