threaddbg.py: be more sensitive about reading variable which are not protected from thread locks.
gdb.py.in: break out list command into a parsing part and a listing
part so that applications which embed the debugger can use the parsing
part but redo the listing part.

fns.py: add this cool function that takes a function body from one
place and a docstring from another. Useful in decorating functions
which want to keep the docstring from the thing they are
decorating. The help mechanism uses docstrings for example. Not used
currently (but may be in ipython).