Doc/* correct email address
pydb.py.in, pydbbdb.py: move OS idiocycracies of set_trace to bdb so as
to make things uniform from the outside.

sighandler.py:  SighanderManager No longer requires pydb object to be passed it
Fix a bug in checking if the right install signal handler is installed. Install
the right one when calling action. Deal with stopping in the debugger even if
we hadn't called the debugger previously. Some of this last stuff
seems a little hacky and could stand improvement.

Whew.